SelectedProduct View Model created
Quantity selection by customer of the store sample is dealt with through
the implementation of a SelectedProductModel, which obtains details
about the selected item from the CatalogEntry on which it is based, and
has an added Quantity property. Quantity is posted back based on user
selection, and a purchase record and associated order items are created.